commit for add new class library persistence and use result pattenrt in the cretae product cmd

+++ src/Core/eInventory.Application/Features/Categories/Handler/Commands/DeleteProductCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using eInventory.Application.Common;
 using eInventory.Application.Features.Categories.Request.Commands;
 using eInventory.Application.Persistence.Contracts;
 using MediatR;
@@ -6,19 +7,27 @@ using MediatR;
 namespace eInventory.Application.Features.Categories.Handler.Commands;
 
 public class DeleteProductCommandHandler(IProductRepository repository, IMapper mapper)
-    : IRequestHandler<DeleteProductCommand>
+    : IRequestHandler<DeleteProductCommand, Result<Unit>>
 {
     public IProductRepository _repository { get; } = repository;
     public IMapper _mapper { get; } = mapper;
 
-    public async Task Handle(DeleteProductCommand request, CancellationToken cancellationToken)
+    public async Task<Result<Unit>> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
     {
         var product = await _repository
             .GetAsync(request.Id, cancellationToken)
             .ConfigureAwait(false);
 
+        if (product is null)
+        {
+            var errors = new List<string>() { $"Product with ID {request.Id} was not found." };
+            return Result<Unit>.Failure(errors, "Delete Failed");
+        }
+
         await _repository
             .DeleteAsync(product, cancellationToken)
             .ConfigureAwait(false);
+
+        return Result<Unit>.Success(Unit.Value, "Delete Successfully");
     }
 }
